+++ src/test/kotlin/com/github/hechtcarmel/jetbrainsindexmcpplugin/server/JsonRpcHandlerUnitTest.kt
@@ -65,22 +65,6 @@ class JsonRpcHandlerUnitTest : TestCase() {
         )
     }
 
-    fun testToolsListRequest() = runBlocking {
-        val request = JsonRpcRequest(
-            id = JsonPrimitive(2),
-            method = JsonRpcMethods.TOOLS_LIST
-        )
-
-        val responseJson = handler.handleRequest(json.encodeToString(JsonRpcRequest.serializer(), request))
-        val response = json.decodeFromString<JsonRpcResponse>(responseJson!!)
-
-        assertNull("${JsonRpcMethods.TOOLS_LIST} should not return error", response.error)
-        assertNotNull("${JsonRpcMethods.TOOLS_LIST} should return result", response.result)
-
-        val result = response.result!!.jsonObject
-        assertNotNull("Result should contain tools array", result["tools"])
-    }
-
     fun testPingRequest() = runBlocking {
         val request = JsonRpcRequest(
             id = JsonPrimitive(4),